rename trans.ts -> i18n.ts, it could get locale functions in future
